# event-manager/src/main/java/dev/sorokin/eventmanager/dto/ErrorMessageResponse.java
package dev.sorokin.eventmanager.dto;

import java.time.LocalDateTime;

public record ErrorMessageResponse(
        String message,
        String detailedMessage,
        LocalDateTime localDateTime
) {
}